Added USB Splitter
When I was designing the headset, I forgot to add the USB Splitter cable - or at least a mount for it. It should now slot in nicely. I also tweaked the tolerances on the Tribar holes again so that you won't need to sand down the tribars to get everything to fit.